Navigation Implementation 1.0
This update gives every Entity both an AR navigation button and a Google Maps navigation button. Upon starting Navigation, if current activity isn't ArActivity, the current activity finishes. The way AR directions are given is currently via a direct vector between the user and the destination. Ideally, this would be split into a series of walkable waypoints so that the user is not trying to be taken through any walls. We may also want to use an AR arrow model instead of something that says "This way" and/or a compass UI at top of screen like Skyrim with the destination highlighted in the compass.

Changes per file:

AndroidManifest.xml:
- All activities are made to be portrait.

AnchorHelper:
- Anchor creation now happens for destination even if outside of proximity distance

ArActivity:
- Included a button to stop navigation

BuildingDialogFragment and LandmarkDialogFragment:
- Refactored and dropped Data from both names
- Included the associated entity in the constructor
- Included AR navigation and Google Maps navigation

Entity, BuildingEntity, and LandmarkEntity:
- Separated dialog creation to init instead of constructor. Needed to use "this" to create dialog and couldn't from constructor
- Included a boolean called isDestination for each entity with ability to get, set, and clear as destination

Navigation:
- This class handles all navigation updates and arrow placement
- There are some companion object methods (so that entity dialog fragments could start navigation as well

SearchActivity:
- Simplified showing an entity's dialog fragment in search activity

activity_ar.xml:
- Added a navigation button

building_dialog.xml and landmark_dialog.xml
- Refactored and dropped data from name
- AR nav button and Map button are included